+++ src/main/java/shit/zen/modules/impl/render/AspectRatio.java
@@ -7,7 +7,7 @@ import shit.zen.settings.impl.NumberSetting;
 public class AspectRatio
 extends Module {
     public static AspectRatio INSTANCE;
-    public final NumberSetting ratioSetting = new NumberSetting("Ratio", Float.valueOf(1.78f), Float.valueOf(0.1f), Float.valueOf(5.0f), Float.valueOf(0.1f));
+    public final NumberSetting ratioSetting = new NumberSetting("Ratio", 1.78f, 0.1f, 5.0f, 0.1f);
 
     public AspectRatio() {
         super("AspectRatio", Category.RENDER);
